% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayMDSApp.R
\name{bayMDSApp}
\alias{bayMDSApp}
\title{Shiny App for exploring the results of \code{bmds} function}
\usage{
bayMDSApp(out)
}
\arguments{
\item{out}{an object of class \code{bmds}, the output of the \code{bmds} function}
}
\value{
open Shiny app
}
\description{
Call Shiny to show the results of Bayesian analysis of
multidimensional scaling in a web-based application.
}
\examples{
\donttest{
data(cityDIST)
out <- bmds(cityDIST, min_p=1, max_p=6 )
if(interactive()){bayMDSApp(out)}
}
}
