% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDIST.R
\name{checkDIST}
\alias{checkDIST}
\title{check the dissimilarity matrix}
\usage{
checkDIST(dist, ...)
}
\arguments{
\item{dist}{dissimilarity measures for pairs of objects}

\item{...}{arguments to be passed to methods}
}
\value{
a full matrix of dissimilarity measures
}
\description{
check the type of dissimilarity matrix and convert it to a symmetric full matrix for the input of \code{bmdsMCMC} and \code{bmds} function
}
\examples{
x <- matrix(rnorm(100), nrow = 5)
dist(x)
checkDIST(dist(x))
}
