\name{cityDIST}
\alias{cityDIST}
\docType{data}
\title{
Airline distances between cities
}
\description{
Airline distances between 30 principal cities of the world. Cities are located
on the surface of the earth, a three-dimensional sphere, and airplanes travel on the surface of the earth.
}

\references{ Hartigan, J.A. (1975), Clustering Algorithms, Wiley, New York.}

\examples{
data(cityDIST)
}
\keyword{datasets}
