% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{distRcpp}
\alias{distRcpp}
\title{calculate Euclidean distances}
\usage{
distRcpp(X)
}
\arguments{
\item{X}{data matrix}
}
\value{
distance matrix
}
\description{
calculate Euclidean distances between rows of matrix X
}
\examples{
x <- matrix(rnorm(100), nrow = 5)
distRcpp(x)

}
