% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDelDist.R
\name{plotDelDist}
\alias{plotDelDist}
\title{plot Delta  vs DIST}
\usage{
plotDelDist(out)
}
\arguments{
\item{out}{the output of the function \code{bmdsMCMC}}
}
\value{
plot of delta vs. dist
}
\description{
plot Delta (estimated Euclidean distance from \code{bmds}) vs DIST (observed dissimilarity measure)
 for pairs of objects
}
\examples{
data(cityDIST)
result <- bmdsMCMC(cityDIST,p=3,nwarm=500,niter=3000)
plotDelDist(result)



}
